Added react test action
