+++ .github/actions/prepare.yml
@@ -1,0 +1,15 @@
+name: Prepare PowerAuth JS environment
+
+runs:
+  using: composite
+  steps:
+      - name: Use Node.js 22
+        uses: actions/setup-node@v4
+        with:
+          node-version: 22
+      - name: Install dependencies
+        run: npm i
+      - name: Set .env file
+        env:
+          ENV_TEST_FILE: ${{ secrets.ENV_TEST_FILE }}
+        run: echo -e "$ENV_TEST_FILE" > testapp/.env
