+++ package.json
@@ -8,7 +8,8 @@
   "source": "src/index",
   "scripts": {
     "build": "node node_modules/gulp/bin/gulp.js",
-    "buildAndRunCordovaIosTests": "npm run build && pushd testapp-cordova && npm run reinstallPlugin && npm run build && cordova run ios && popd && node test-listener.js"
+    "buildAndRunCordovaIosTests": "npm run build && pushd testapp-cordova && npm run reinstallPlugin && npm run build && npm run ios && popd && node test-listener.js",
+    "buildAndRunReactIosTests": "npm run build && pushd testapp && npm run reinstallPlugin && npm run ios && popd && node test-listener.js"
   },
   "files": [
     "README.md",
