+++ testapp-cordova/package.json
@@ -5,9 +5,9 @@
     "description": "A sample Cordova application that runs PowerAuth tests.",
     "main": "index.js",
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "installPlugin": "cordova plugin add ../build/cdv",
-        "reinstallPlugin": "(cordova plugin remove cordova-powerauth-mobile-sdk || true) && npm run installPlugin",
+        "ios": "npx cordova run ios",
+        "installPlugin": "npx cordova plugin add ../build/cdv",
+        "reinstallPlugin": "(npx cordova plugin remove cordova-powerauth-mobile-sdk || true) && npm run installPlugin",
         "build": "node node_modules/gulp/bin/gulp.js"
     },
     "keywords": [
